Update multiSearch to not require too much
The caller now can just pass useful info to the method. They should not
specify the `json` key (it's done internally), and they should not
`json_encode` it either (also done internally). Finally, within a
`request` the `meta` array is converted to the form-encoded version.

Note: it's still up to the caller to encode filters and scales per
normal.

See `examples/1/multi_search.php` for usage.